Design MainActivity - in progress
