+++ app/src/main/AndroidManifest.xml
@@ -11,6 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.BeeSpotter"
         tools:targetApi="31">
+
         <activity
             android:name=".MainActivity"
             android:exported="true">
